Moving components to appropriate folders
