+++ src/components/GoalsList/GoalCard/GoalCard.styles.js
