+++ src/components/GoalsList/GoalsList.js
@@ -3,7 +3,7 @@ import { UserContext } from '../../context/UserContext';
 import Grid from '@material-ui/core/Grid';
 import Button from '@material-ui/core/Button';
 import Typography from '@material-ui/core/Typography';
-import GoalCard from './GoalCard';
+import GoalCard from './GoalCard/GoalCard';
 import axios from 'axios';
 
 const GoalsList = ({ setLoading, setAlert }) => {
