+++ src/components/LearningDaysList/LearningDayCard/LearningDayCard.styles.js
