+++ src/components/LearningDaysList/LearningDaysList.js
@@ -4,7 +4,7 @@ import Menu from '../../components/Menu/Menu';
 import Grid from '@material-ui/core/Grid';
 import Button from '@material-ui/core/Button';
 import Typography from '@material-ui/core/Typography';
-import LearningDayCard from './LearningDayCard';
+import LearningDayCard from './LearningDayCard/LearningDayCard';
 import axios from 'axios';
 
 const LearningDaysList = ({ setLoading, setAlert }) => {
